# By Dylan Graham (40) and others
# Via Dylan Graham (6) and others
* 'master' of https://github.com/dylangraham/dna: (54 commits)
  Fix batch form within form
  Cleanup SourceTree crap
  Styled batched form
  Remove link on batch ID
  Add user as super to test nav
  Clean up session each time sample edited
  removed unused import in provider
  user check of is super nav bar and batch pages
  Made Batch form responsive
  Remove i5 value if not selected
  Remove logo
  "removed extra js call in batchrun
  fixed back button on batchruns form
  Create run with form input
  updated validations, set default run_buit hidden
  updated validation regex to anchor start and end
  update several validations adding unique value
  rolled back batch_check_id appending id in runs form
  Initial Vagrantfile
  Add historical sample seeds
  ...

+++ database/migrations/2015_08_15_116124_create_batches_table.php
@@ -22,7 +22,7 @@ class CreateBatchesTable extends Migration
             $table->string('tube_bar_code', 60);
             $table->string('tube_location', 60);
 
-            $table->float('tape_station_length');
+            $table->integer('tape_station_length');
             $table->string('charge_code', 17);
             $table->timestamps();
